Add support for using JBang (#575)
* Fix typo and use GAV that exist in maven central

* Add support for using JBang

Looks for `jbang.properties` as a marker similar to `pom.xml` and `settings.gradle`.
It is not required for jbang to work, thus we could also just look for a standard source file i.e. `src/main.java`

Also handles `binary: myFile.java` as a way to customize which of possible "scripts" to use.

Also allow using Groovy (.groovy) and Kotlin (.kt) via JBang, but that is untested for now.

* Add jbang example to the test suite

* Add support for using JBang

Looks for `jbang.properties` as a marker similar to `pom.xml` and `settings.gradle`.
It is not required for jbang to work, thus we could also just look for a standard source file i.e. `src/main.java`

Also handles `binary: myFile.java` as a way to customize which of possible "scripts" to use.

Also allow using Groovy (.groovy) and Kotlin (.kt) via JBang, but that is untested for now.

* apply PR suggestions

* Satisfy lint, split jbang.go

* Lint

Co-authored-by: Anton Tayanovskyy <anton@pulumi.com>