Lint depending on pkg except for codegen
To avoid including all pkg deps in tf-bridge, we want to move codegen
into its own module. Since codegen depends on pulumi-java, we need to
make sure that pulumi-java doesn't pull in the rest of pulumi/pkg. We
can enforce this with a lint rule.

The rest of this PR is getting the linter happy.