sdk/StackReference: Add outputDetailsAsync
Adds the Java equivalent of StackReference.GetOutputDetails
that was added to:

- Go SDK in pulumi/pulumi#12034
- Python SDK in pulumi/pulumi#12071
- Node SDK in pulumi/pulumi#12072
- .NET SDK in pulumi/pulumi-dotnet#103

In Java, this adds a new `StackReference.outputDetailsAsync` method
which returns a CompletableFuture of a StackReferenceOutputDetails.
The object has optional value and secretValue fields,
at most one of which is set depending on whether the value is a secret.
Neither field is set if the value is absent.

Refs pulumi/pulumi#10839, pulumi/pulumi#5035