ci/go-tests: Increase timeout
The timeout for the go-tests job is 30 minutes
but the tests are allowed only 15 to run.
Looking at recent job successes, go-tests takes very close to 15 minutes to run.

Most recent failures have been because this operation timed out
by hitting the 15 minute limit.

Increase the timeout for this task to 25 minutes.