Remove changes to minimal example
Got PR feedback requesting removing the fixes to the minimal example.

Anton: In that case this might belong in a separate PR? Is it failing on master or why did it start failing with the changes? The minimal example test is doing double-duty as a canary for misaligned dependencies, and I'm worried this change papers over that. Misaligned dependencies sometimes result in funny breakage for users depending on their maven import order of pulumi dips.