Update `upload-artifact` and `download-artifact` actions (#1383)
v3 of `actions/upload-artifact` and `actions/download-artifact` will be
deprecated on November 30, 2024. This commit bumps the versions of these
actions to v4. While there are breaking changes between v3 and v4
(namely, the introduction of immutable artifacts), it doesn't appear
that `pulumi-java` falls foul of them, so this is hopefully as simple as
bumping the version.