Fully qualify `java.lang` types in codegen (#1400)
When generating code, it's entirely possible for us to encounter
resources and functions with names that clash with those found in the
Java standard library, such as `String` and `Integer`. We must therefore
be careful to qualify standard names so that they are unambiguous in
generated code. This commit fixes up a few gaps to this effect, ensuring
that we emit e.g. `java.lang.String` and not just `String` when
referring to Java's own types.