Upgrade golangci-lint to the latest version
CI workflows are currently configured to use the latest stable version of Go, which is now v1.23. With this new version of Go, we're seeing new lint errors. Using the latest version of golangci-lint addresses the vast majority of lint failures, aside from two cases where a `nil` check is unnecessary, which this change also addresses.