Exclude Non-Loadable Packages from ClassLoader. Fixes #issue 1385 (#1384)
We encountered an issue with the package loader that loads every package
indiscriminately, instead of using the Maven class loader's smart
loading. This is a known problem, and while some packages have already
been excluded, the process requires a merge request for each exclusion.
Therefore, we identified the need for a faster and more dynamic
exclusion mechanism.

This fix introduces a system property string containing a
comma-separated list of packages to exclude. By setting this property,
the specified packages will be excluded from the loaded packages.

Fixes #1385