implement debugging support for Java
For Java to run the program with debugger support, we currently only
support maven, so we can use mvnDebug to run it. There might be other
ways to start the debugger using
https://github.com/microsoft/java-debug, which could be more generic,
but the documentation is somewhat lacking, and this works.  There's
always a possibility of swapping this out for java-debug later if we
find this method lacking.