Feature flag parameterized provider support
Support for parameterized providers (introduced originally in #1398) is now
exposed by the engine as a new `parameterized` feature flag. This commit updates
the Java SDK to check for support before attempting to register a parameterized
package, allowing us to throw a better error message in the event that
parameterization is not currently supported.

Part of pulumi/pulumi#16906