pin the version of examples and templates
The tests for pulumi-java rely on tests in the examples and templates
repo.  However those repos are moving targets, where tests can get
deleted, changed etc., which can break tests here.  This is obviously
a non-ideal situation.  For now just pin the version of the examples
and templates repo, so this will always keep working.