Improve error messages when Java commands fail
When we fail to execute a Java command, we currently duplicate the command name
(e.g. `mvn`), which makes it harder to understand what has gone wrong. This
commit cleans this up, additionally quoting the command we attempted to run so
that it's easier to parse in the event it contains spaces.