Modernise root resource handling
The root resource of a stack is the top-level parent of all resources created in
that stack. Historically, we constructed a URN for this root resource by making
a call to a `GetRootResource` endpoint on the engine. This is no longer the case
however, and other language SDKs now handle this directly by using the stack's
URN as the default root resource. This commit makes the analogous change for the
Java SDK.