Implement `GenerateProgram` and `GenerateProject`
The `GenerateProgram` and `GenerateProject` methods of the `LanguageRuntime`
gRPC service allow callers to generate programs (in this case, Java source code)
and projects (programs complete with e.g. build and packaging components -- in
this case, Maven `pom.xml`s and so on). This commit implements these endpoints
for `pulumi-language-java`, using the functionality that already exists but
previously was only exposed through the `pulumi-java-gen` binary. Doing so means
we can begin to get conformance tests (which generate programs using RPC calls)
working for Java.

Note that as part of this, we add support for local (packed) dependencies to
`GenerateProject`, which is key to enabling conformance tests also. We do this
by harnessing Maven's existing `<repositories>` tag, which allows specifying
paths to non-default Maven repositories where it can find `.jar`s etc. This will
tie in with our later implementation of `Pack`, which will produce local Maven
repositories containing built `.jar`s.