Implement `GeneratePackage` (#1470)
The `GeneratePackage` method of the `LanguageRuntime` gRPC service
allows callers to generate SDKs for a specified schema. This commits
implements this endpoint for `pulumi-language-java`, using the existing
`pkg/codegen` functionality that is currently only exposed through the
`pulumi-java-gen` binary. Doing so means we can begin to get conformance
tests working for Java.

As part of this work, we extend the work we started when implementing
`GenerateProject` to support local repositories in generated SDKs. As
with other languages, we add a field to `PackageInfo` to this end. Since
SDK generation uses Gradle (and not Maven, as program generation does),
there is a bit of new plumbing to do here, but in principle the outcome
is the same.

> [!NOTE]
> It is not certain yet that `GeneratePackage` behaves identically to
the currently-used binary endpoint in `pulumi-java-gen` -- we should
make sure of this before switching over. This work is primarily a means
to get conformance tests going.