Implement conformance tests for Java (#1403)
This commit pieces together the newly added `GenerateProgram`,
`GenerateProject`, `GeneratePackage` and `Pack` calls to enable language
conformance tests for Java, initially running just the `l1-empty` test.
We take care of the following in order to make this happen:

* Add the fairly standard `language_test.go`/`TestLanguage` boilerplate
for kicking off the test suite against a Java language host. As part of
this we take care of some Java-specific hacks, such as copying `.proto`
files for the core SDK from the `pulumi` submodule, and excluding tests
that don't pass for now.
* Patch up the `Run` and `InstallDependencies` RPC endpoints to respect
the supplied `ProgramDirectory` when executing Java commands.
* Implement `GetProgramDependencies` for Maven-executed Java programs.
This seems sufficient for now since program generation only generates
Maven-built code. We need to implement `GetProgramDependencies` in order
for the conformance tests to pass, since they validate the list of
dependencies against an expected list of generated provider SDKs, for
instance. We omit versions when enumerating dependencies for now, since
there are some discrepancies between those we report and those that are
expected.

Part of https://github.com/pulumi/pulumi/issues/17505