Remove use of BaseProvider.PluginDownloadURL (#1483)
We duplicated PluginDownloadURL into the BaseProvider schema block, but
this isn't needed. The PluginDownloadURL field in the top of the schema
will always have the same information, and it only makes sense to talk
about _one_ download url for the package because there's only one plugin
to download.

This starts the process of removing the field from BaseProvider, once
this is merged we can pull this update to pu/pu and remove the field
completely from there.