Explicitly install pulumi CLI in CI (#1487)
We were relying on the Pulumi version that comes preinstalled on GHA
runners. This got updated to 3.141.0 and we started hitting
https://github.com/pulumi/pulumi/issues/17862 in our template tests.

Explicitly install the latest version, so once we release a fix for the
issue we can pass CI without waiting for a new runner image.

Tests will fail until 3.142.0 is out.