Implement `GetRequiredPackages` for Java
As part of a Pulumi deployment, the Pulumi engine will ask the language host
running the program for the set of plugins required to run that program
successfully by making a `GetRequiredPlugins` gRPC call. For instance, if one
has a TypeScript program whose `index.ts` `import`s `@pulumi/aws`, the NodeJS
language host will find the `@pulumi/aws` package and report that the `aws`
resource plugin is required. This enables features such as pre-emptively
downloading a plugin ahead of time, as opposed to waiting for the first resource
registration against that plugin.

With the introduction of parameterized providers, `GetRequiredPlugins` falls a
little short -- parameterization means that it is no longer necessarily the case
that a package named `x` is provided by a plugin of the same name. For instance,
with a dynamically bridged Terraform provider X, the provider name will be X,
but the plugin name will be e.g. `pulumi-terraform-provider`. In
pulumi/pulumi#17894, `GetRequiredPackages` was added to the language host
interface to provide the additional context of a plugin's parameterization, and
is now used if present instead of `GetRequiredPlugins`. This change implements
`GetRequiredPackages` for Java as follows:

* Presently, `GetRequiredPlugins` is implemented by running a small Java program
  (located in the internal `bootstrap` package of the core SDK) that walks its
  class path looking for `pulumi-plugin.json` files. This code is extended to
  parse `parameterization` blocks from those JSON files, as we do in other
  languages using this mechanism.
* When we generate SDKs, we generate Gradle build files which generate
  `pulumi-plugin.json` files as part of the build process. We extend the
  `build.gradle` template used when generating SDK build files so that it emits
  `parameterization` blocks appropriately. As part of this, we also clean up the
  emission of package names and versions.
* We now have enough to run and pass the `l2-parameterized-resource` conformance
  test, which in theory confirms that all of this works, so we include that as
  part of this work.
* We remove the now-deprecated `GetRequiredPlugins` implementation.

Fixes #1510