Support exporting plain stack output values
This change adds an overload of `Context.export` that accepts plain values that
aren't `Output<T>` so that they can be exported without having to be manually
wrapped in `Output.of(...)`, making user code more succinct and ergonomic. The
new overload is implemented as a default method that wraps the passed-in value
with `Output.of(...)` and then passes the result to the existing overload. This
aligns with how args can be passed to resources either as a plain value or an
`Output<T>`, via overloads with the builder pattern.

An additional benefit is that it allows us to enable more conformance tests, as
programgen currently only passes plain values to `Context.export` (it doesn't
currently support wrapping the value with `Output.of`).