[program-gen] Fix generation of `double` literals in generated programs (#1580)
### Description

This PR implements an initial pass for handling `__convert` calls in
program-gen starting with `double` literals. Later on we will extend
this to handle enums but we don't have a conformance test for these yet.

The return type of `__convert` calls is annotated with the target schema
type so we detect when are going to emit an expression of type
`model.NumberType` (in PCL space) in a location that expects
`schema.NumberType`. When that pattern is detected we make sure to emit
double literals that end with `.0` when the value of the number is
exact.

- Fixes #1565 
- Unskips `l2-resource-primitive` now that the test is passing
- ~~Removes `l2-map-keys` as there is no such test~~