[program-gen]  Fix generation of untyped maps and array literals (#1581)
### Description

Fixes how we are generating map and array literals when they are
_untyped_. The problem was that we generated just the elements i.e. `[1,
2, 3]` in PCL would generate `1, 2, 3` in java because Array-typed
resource inputs accepts the elements being spread like that i.e.
`resourceInputs = [1,2,3]` (PCL) generated `resourceInputs(1, 2, 3)`
(Java). However this doesn't work when we are initializing lists and
maps outside of resource inputs: local variables and output variables.

This PR addresses the problem by using `Map.ofEntries(...)` and
`List.of(...)` when applicable.

Branch is written on top of #1580

- Fixes #1561 
- Fixes #1560 
- Unskips `l1-output-map`
- Unskips `l1-output-array`