Remove `@Deprecated` Java SDK members
As part of the 1.0.0 release, this change removes parts of the Java SDK that
have been marked as deprecated and are no longer in use by internal Pulumi
functions. The `Config.of` method has been left as it looks like we might still
be generating code that uses it -- this will be addressed in a future piece of
work if so.

> [!NOTE]
> This is incompatible with some old SDKs that use annotations or annotation
> fields that are no longer present (or searched for), such as the `type` and
> `parameters` fields of the `Export` annotation. Some rough checking seems that
> we made this change a few years ago, so it should be relatively safe, but it is
> still breaking nonetheless (as is to be expected, arguably).