[sdk] Implement InvokeOptionsBuilder and InvokeOutputOptionsBuilder (#1585)
### Description

The PR implements the following builder classes
 - `InvokeOptionsBuilder` for `InvokeOptions`
 - `InvokeOutputOptionsBuilder` for `InvokeOutputOptions`

It adds `pluginDownloadURL` for both of them.

This is a required in order to invoke options for program-gen for #1566

### API

Example usage for `InvokeOptionsBuilder`
```java
InvokeOptions.builder() // returns InvokeOptionsBuilder
  .parent(...)
  .provider(...)
  .pluginDownloadURL(...)
  .version(...)
  .build() // returns InvokeOptions
```

Example usage for `InvokeOutputOptionsBuilder`

```java
(new InvokeOutputOptionsBuilder())
  .parent(...)
  .provider(...)
  .pluginDownloadURL(...)
  .version(...)
  .dependsOn(...) 
  .build() // returns InvokeOutputOptions
```

The reason `InvokeOutputOptionsBuilder` cannot be initialized via
`InvokeOutputOptions.builder()` is because the `builder()` method would
conflict with the `InvokeOptions.builder()` method.

For the time being, we are omitting `.builder()` until we no longer have
`InvokeOutputOptions` _extending_ `InvokeOptions.`