Default to using gradle-nexus in gen-sdk (#1612)
When using `pulumi-java-gen`, the user can bass a `—build` flag to set
gradle related settings in the PackageInfo:

| --build | PackageInfo | Behaviour |

|--------------------|----------------------|--------------------------------------|
|**`none`** | | no gradle file, default |
|`gradle ` | `buildFiles: gradle` | gradle file without nexus plugin |
|`gradle-nexus` | `buildFiles: gradle, gradleNexusPublishPluginVersion:
2.0.0` | gradle file with nexus plugin, using default version of the
plugin |
|`gradle-nexus:$VER` | `buildFiles: gradle,
gradleNexusPublishPluginVersion: $VER` | gradle file with nexus plugin,
using specified version of the plugin |

Pulumi providers use `--build gradle-nexus`.

When using `gen-sdk` instead of `pulumi-java-gen`, you control the
gradle generation with the `PackageInfo`. To streamline the usage of the
most desired behaviour (`gradle-nexus`), this PR changes the behaviour
of the `buildFile` option when using `gen-sdk`. The behaviour when using
`pulumi-java-gen` is unchanged.

This allows providers to not set any options in their schema and still
get the best default behaviour.
 
| PackageInfo | Behaviour |

|--------------------------------|--------------------------------------|
| `buildFiles: none` | no gradle file |
| **`buildFiles: gradle-nexus`** | gradle file with nexus plugin, using
default version of the plugin, `2.0.0`, the default |
| `buildFiles: gradle-nexus, gradleNexusPublishPluginVersion: $VER` |
gradle file with nexus plugin, using version $VER of the plugin |
| `buildFiles: gradle` | gradle file without nexus plugin |
| `buildFiles: gradle, gradleNexusPublishPluginVersion: $VER` | gradle
file with nexus plugin, using version $VER of the plugin |

Fixes https://github.com/pulumi/pulumi-java/issues/1593