auto: Initial interfaces and supporting types (#1618)
This adds the abstract `Workspace` class and associated supporting
classes for the Java Automation API. Currently, these are going under
`com.pulumi.experimental.automation`, but the intention is to move
everything to `pulumi.com.automation` once all the support has landed
and is stable.

Next up will be `LocalWorkspace` and `WorkspaceStack`, and supporting
code, which contains the most of the implementation, along with tests.