Inline runner machinery (#1616)
The use case is to be able to create resources without an arbitrary
function that is created inside a provider (`Construct` in particular).
All methods are marked internal and are intended to be consumed from our
own SDK.

Based on our .NET implementation in
https://github.com/pulumi/pulumi-dotnet/tree/main/sdk/Pulumi/Deployment

Towards https://github.com/pulumi/pulumi-java/issues/1335

No testing yet - but opening it up also for @justinvp to take a look in
case there may be any overlap with Automation API work.

---------

Co-authored-by: Justin Van Patten <jvp@justinvp.com>