Remove `@Deprecated` Java SDK members (#1587)
As part of the 1.0.0 release, this change removes parts of the Java SDK
that have been marked as deprecated and are no longer in use by internal
Pulumi functions. The `Config.of` method has been left as it looks like
we might still be generating code that uses it -- this will be addressed
in a future piece of work if so.