Configure and Construct implementation in the provider service (#1645)
This PR implements Configure and Construct in the provider protocol -
enough to enable very basic Component Packages. Serializing and
deserializing to/from PropertyValues has to be done by component
authors. (utilities for that are coming in later PRs)

Both Maven and Gradle provider tests are changed from a basic provider
returning a static schema to a functional Component provider with a
basic component. Tests deploy those with ProgramTest and verify the
output shape.