Add Automation API (#1657)
This change adds Automation API to `com.pulumi.automation` and prepares for the v1.3.0 release.

A small change is also included that renames some static fields from `EMPTY` to `Empty` to match the casing we've already used elsewhere in the core SDK and generated provider SDKs. (It should be its own separate change, but including it here for expediency).

Fixes #1633