Add ability to disable toString on Output<T>
This change provides an ability that is already present in the Python and Node.js SDK, and recently added to the .NET SDK, to to disable converting `Output<T>` to a string.

Currently in the Java SDK this will return a helpful message in the string contents to advise not to do this but does not fail the Pulumi program.

This functionality is opted-in by setting the environment variable `PULUMI_ERROR_OUTPUT_STRING` to `true` or `1` to throw an exception.