Update pulumi submodule to master and use the submodule for pulumi-java
This way for testing on pull-recursive you always have the latest
submitted pulumi/pulumi for tests and you can run against it instead of
pulling a released version.

The logic here is we develop against head and have to manually do this,
so why not leave the repository in that state and change the release
process to check out a specific SDK if necessary?