Dev/davide/context aware fixup (#1678)
The current code was relying on a static field to track the Deployment
context.
This meant that it was not possible to run parallel deployments.

These changes swap the static field with thread-local storage. Because
the system heavily depends on asynchronous operations, and the runtime
controls which thread executes what callback, we need to inject the
correct state into the thread at each reentrancy point.

ContextAwareCompletableFuture<T> is a subclass of CompletableFuture<T>
which exposes the same helper methods, but captures and reapplies the
deployment state on restart.

---------

Co-authored-by: Mikhail Shilkov <github@mikhail.io>