Add default `ci-ok` checks for `CHANGELOG`-only PRs (#1730)
When a pull request (PR) only touches `CHANGELOG*` files, we skip CI
checks in order to make processes that only modify these files, such as
releases, much faster. Recently, we tightened controls on this
repository so that all PRs require a passing `ci-ok` step in order to
merge. As a result of this, we need to tweak our skipping approach to
produce a "no-op" `ci-ok` step so that `CHANGELOG*`-only PRs can merge.
This change does that by copying the GitHub-sanctioned pattern used in
`pulumi/pulumi`. This will be tested by the upcoming 1.6.0 release PR.