Default version 0.0.0 for component providers
We need a version to code generate an SDK that can be used right away, without modifying things like `package.json` or `pyproject.toml` with the version after the fact.

Always return 0.0.0 for component providers. The engine will take of overriding the version for plugins loaded from GItHub and set the correct version.