run CI tests on a regular schedule (#1767)
Run the CI tests in this repo on a regular schedule, to guard against
bitrot. "Inspired" by
https://github.com/pulumi/pulumi-java/actions/runs/14494159872/job/40657808628?pr=1766
failing, which I don't think has anything to do with my PR, but if we
had a cron job running the tests on a regular schedule it would be
easier to check.

Also rename the CI job from pr.yml -> ci.yml, since it's not only
running on PRs already and even less so now.