Implement `Language.Link` (#1915)
This takes the instructions from
https://github.com/pulumi/pulumi/blob/96d8ca5799400653b31926cf3c1574e174c0f46e/pkg/cmd/pulumi/packages/packages.go#L444-L468
and moves them to the `Link` RPC. This does not perform any modification
of project files, it merely prints instructions the user has to follow
to link the generated SDK.

Note that link can potentially link multiple packages at once.

Once this is merged and pulumi-java is updated in the main pulumi repo,
we can drop the `printJavaLinkInstructions` function from there.

Closes https://github.com/pulumi/home/issues/4373