Manage changelogs with `changie` (#1928)
Java's deploy process is just a little bit more effort than dotnet and
yaml. We have to merge the PENDING entries into CHANGELOG, which
triggers a job that uses the content of PENDING, and then push some tags
and finally remove PENDING. With `changie`, we can always see what would
have been the contents of PENDING using the `latest` command, so we can
remove the final step entirely: do a `changie` PR as we would in dotnet
or java, and then push the tags and we're done.

In a follow-up, I'd like to take out the second step there, but this one
was already looking scary. Really, this change should be +44/-7 but we
have to migrate all previous entries and reformat the changelog, so it
looks much scarier than it is.